pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docslice"
version = "0.1.0"
description = "Slice documentation sites into clean Markdown files for LLM workflows."
readme = "PRD.md"
requires-python = ">=3.10"
dependencies = [
  "beautifulsoup4>=4.12",
  "httpx>=0.27",
  "lxml>=5.2",
  "markdownify>=0.13",
  "pydantic>=2.7",
  "python-slugify>=8.0",
  "PyYAML>=6.0",
  "typer>=0.12",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
]

[project.scripts]
docslice = "docslice.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
